Batch Prettier changes
When Prettier formats a file, especially when it formats the file for
the first time, doing so may result in a large number of small changes
to the buffer.

The improvement in this commit uses heuristics and the
combine-change-calls macro in an attempt to reduce the number of hook
calls that may happen in this scenario.

The relevant Emacs documentation at
https://www.gnu.org/software/emacs/manual/html_node/elisp/Change-Hooks.html
says about this that

> [the combine-change-calls] macro is useful when a function makes a
> possibly large number of repetitive changes to the buffer, and the
> change hooks would otherwise take a long time to run, were they to
> be run for each individual buffer modification. Emacs itself uses
> this macro, for example, in the commands comment-region and
> uncomment-region.

And in the related documentation for combine-after-change-calls:

> If a program makes several text changes in the same area of the
> buffer, using the macro combine-after-change-calls around that part
> of the program can make it run considerably faster when after-change
> hooks are in use.