Change process connection type to pipe
It turns out that the IPC between prettier.el and prettier-el.js was
somehow broken on Emacs on WSL1. Setting `process-connection-type` to
`nil` so that `new-process` uses a pipe instead of a pty (according to
the Emacs documentation linked below) fixes the behavior.

Disclaimer: I haven't tested this fix on proper Linux or any other
platforms than WSL.

https://www.gnu.org/software/emacs/manual/html_node/elisp/Asynchronous-Processes.html